Added the refractor files
Makes file sizes smaller as well as staying more organized
